fix(benchmarks): move account_query test from compute to stateful
